+++ release/ui/layout/BadgeFooter/BadgeFooter.scss.es6.js
@@ -1,7 +1,7 @@
 import injector_8363bc92 from 'rollup-plugin-styles/dist/runtime/inject-css.js';
 
-var css = ".BadgeFooter_root__1ee58a84{background:var(--color-primary-50);border-bottom-left-radius:inherit;border-bottom-right-radius:inherit;bottom:var(--ui-badge-footer-offset,0);color:var(--color-100);left:var(--ui-badge-footer-offset,0);position:absolute;right:var(--ui-badge-footer-offset,0)}";
-var modules_b78b788d = {"root":"BadgeFooter_root__1ee58a84"};
+var css = ".BadgeFooter_root__8f5a0fbd{background:var(--color-primary-50);border-bottom-left-radius:inherit;border-bottom-right-radius:inherit;bottom:var(--ui-badge-footer-offset,0);color:var(--color-100);font-size:14px;left:var(--ui-badge-footer-offset,0);position:absolute;right:var(--ui-badge-footer-offset,0)}";
+var modules_b78b788d = {"root":"BadgeFooter_root__8f5a0fbd"};
 injector_8363bc92(css,{});
 
 export { css, modules_b78b788d as default };