+++ release/ui/layout/BadgeFooter/BadgeFooter.scss.js
@@ -8,8 +8,8 @@ function _interopDefaultLegacy (e) { return e && typeof e === 'object' && 'defau
 
 var injector_8363bc92__default = /*#__PURE__*/_interopDefaultLegacy(injector_8363bc92);
 
-var css = ".BadgeFooter_root__1ee58a84{background:var(--color-primary-50);border-bottom-left-radius:inherit;border-bottom-right-radius:inherit;bottom:var(--ui-badge-footer-offset,0);color:var(--color-100);left:var(--ui-badge-footer-offset,0);position:absolute;right:var(--ui-badge-footer-offset,0)}";
-var modules_b78b788d = {"root":"BadgeFooter_root__1ee58a84"};
+var css = ".BadgeFooter_root__8f5a0fbd{background:var(--color-primary-50);border-bottom-left-radius:inherit;border-bottom-right-radius:inherit;bottom:var(--ui-badge-footer-offset,0);color:var(--color-100);font-size:14px;left:var(--ui-badge-footer-offset,0);position:absolute;right:var(--ui-badge-footer-offset,0)}";
+var modules_b78b788d = {"root":"BadgeFooter_root__8f5a0fbd"};
 injector_8363bc92__default["default"](css,{});
 
 exports.css = css;