+++ release/ui/popups/Modal/Modal.d.ts
@@ -1,7 +1,8 @@
-import { type HTMLStyleProps } from '@innet/dom';
+import { type HTMLStyleProps, type StateProp } from '@innet/dom';
 import { type ButtonProps } from '../../buttons';
 export interface ModalProps extends Omit<HTMLStyleProps<HTMLDivElement>, 'onclose'> {
-    width?: number;
+    height?: StateProp<number | string>;
+    width?: StateProp<number | string>;
     buttons?: string[];
     headButtons?: string[];
     buttonProps?: Record<string, ButtonProps>;
@@ -9,4 +10,4 @@ export interface ModalProps extends Omit<HTMLStyleProps<HTMLDivElement>, 'onclos
     onclosed?: (result: string) => void;
     onshow?: () => void;
 }
-export declare function Modal({ buttons, width, style, headButtons, buttonProps, onclosed, onclose, onshow, onmousedown, ...props }?: ModalProps): any;
+export declare function Modal({ buttons, width, height, style, headButtons, buttonProps, onclosed, onclose, onshow, onmousedown, ...props }?: ModalProps): any;